chtml/html/testdata/webkit/blocks.dat

#data
<!doctype html><p>foo<address>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “address”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <address>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><address><p>foo</address>bar
#errors
#document
| <!DOCTYPE html>
| <address>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<article>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “article”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <article>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><article><p>foo</article>bar
#errors
#document
| <!DOCTYPE html>
| <article>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<aside>bar<p>baz
#errors
(1,37): expected-closing-tag-but-got-eof
28: Unclosed element “aside”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <aside>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><aside><p>foo</aside>bar
#errors
#document
| <!DOCTYPE html>
| <aside>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<blockquote>bar<p>baz
#errors
(1,42): expected-closing-tag-but-got-eof
33: Unclosed element “blockquote”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <blockquote>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><blockquote><p>foo</blockquote>bar
#errors
#document
| <!DOCTYPE html>
| <blockquote>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<center>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “center”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <center>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><center><p>foo</center>bar
#errors
#document
| <!DOCTYPE html>
| <center>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<details>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “details”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <details>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><details><p>foo</details>bar
#errors
#document
| <!DOCTYPE html>
| <details>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<dialog>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “dialog”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <dialog>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><dialog><p>foo</dialog>bar
#errors
#document
| <!DOCTYPE html>
| <dialog>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<dir>bar<p>baz
#errors
(1,35): expected-closing-tag-but-got-eof
26: Unclosed element “dir”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <dir>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><dir><p>foo</dir>bar
#errors
#document
| <!DOCTYPE html>
| <dir>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<div>bar<p>baz
#errors
(1,35): expected-closing-tag-but-got-eof
26: Unclosed element “div”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <div>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><div><p>foo</div>bar
#errors
#document
| <!DOCTYPE html>
| <div>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<dl>bar<p>baz
#errors
(1,34): expected-closing-tag-but-got-eof
25: Unclosed element “dl”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <dl>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><dl><p>foo</dl>bar
#errors
#document
| <!DOCTYPE html>
| <dl>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<fieldset>bar<p>baz
#errors
(1,40): expected-closing-tag-but-got-eof
31: Unclosed element “fieldset”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <fieldset>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><fieldset><p>foo</fieldset>bar
#errors
#document
| <!DOCTYPE html>
| <fieldset>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<figcaption>bar<p>baz
#errors
(1,42): expected-closing-tag-but-got-eof
33: Unclosed element “figcaption”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <figcaption>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><figcaption><p>foo</figcaption>bar
#errors
#document
| <!DOCTYPE html>
| <figcaption>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<figure>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “figure”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <figure>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><figure><p>foo</figure>bar
#errors
#document
| <!DOCTYPE html>
| <figure>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<footer>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “footer”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <footer>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><footer><p>foo</footer>bar
#errors
#document
| <!DOCTYPE html>
| <footer>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<header>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “header”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <header>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><header><p>foo</header>bar
#errors
#document
| <!DOCTYPE html>
| <header>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<hgroup>bar<p>baz
#errors
(1,38): expected-closing-tag-but-got-eof
29: Unclosed element “hgroup”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <hgroup>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><hgroup><p>foo</hgroup>bar
#errors
#document
| <!DOCTYPE html>
| <hgroup>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<listing>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “listing”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <listing>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><listing><p>foo</listing>bar
#errors
#document
| <!DOCTYPE html>
| <listing>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<menu>bar<p>baz
#errors
(1,36): expected-closing-tag-but-got-eof
27: Unclosed element “menu”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <menu>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><menu><p>foo</menu>bar
#errors
#document
| <!DOCTYPE html>
| <menu>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<nav>bar<p>baz
#errors
(1,35): expected-closing-tag-but-got-eof
26: Unclosed element “nav”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <nav>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><nav><p>foo</nav>bar
#errors
#document
| <!DOCTYPE html>
| <nav>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<ol>bar<p>baz
#errors
(1,34): expected-closing-tag-but-got-eof
25: Unclosed element “ol”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <ol>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><ol><p>foo</ol>bar
#errors
#document
| <!DOCTYPE html>
| <ol>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<pre>bar<p>baz
#errors
(1,35): expected-closing-tag-but-got-eof
26: Unclosed element “pre”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <pre>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><pre><p>foo</pre>bar
#errors
#document
| <!DOCTYPE html>
| <pre>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<section>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “section”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <section>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><section><p>foo</section>bar
#errors
#document
| <!DOCTYPE html>
| <section>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<summary>bar<p>baz
#errors
(1,39): expected-closing-tag-but-got-eof
30: Unclosed element “summary”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <summary>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><summary><p>foo</summary>bar
#errors
#document
| <!DOCTYPE html>
| <summary>
|   <p>
|     "foo"
| "bar"

#data
<!doctype html><p>foo<ul>bar<p>baz
#errors
(1,34): expected-closing-tag-but-got-eof
25: Unclosed element “ul”.
#document
| <!DOCTYPE html>
| <p>
|   "foo"
| <ul>
|   "bar"
|   <p>
|     "baz"

#data
<!doctype html><ul><p>foo</ul>bar
#errors
#document
| <!DOCTYPE html>
| <ul>
|   <p>
|     "foo"
| "bar"
